Updated naming to fix merge conflict
